requirements.txt
# Core dependencies
Django>=4.2.0
django-bootstrap5>=23.3
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
channels>=4.0.0
Pillow>=10.0.0

# ML dependencies
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch>=2.0.0
transformers>=4.30.0
numpy<2.0.0  # Pinning numpy to 1.x version for compatibility
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
opencv-python>=4.5.0

